balls

+++ src/pruning/taylor_series/final/Rewards.py
@@ -60,7 +60,7 @@ class GetResults:
         self.learndata.eval()
 
         with torch.no_grad():
-            x = torch.tensor([pruning_amount, pre_pruning_acc, initial_acc])
+            x = torch.tensor([pruning_amount, pre_pruning_acc, initial_acc], dtype=torch.float32)
             x = x.unsqueeze(0).to(self.device)
             print(f"Input to model: {x}")
             return self.learndata(x)
